backend/requirements.txt
python-socketio==5.10.0
aiohttp==3.9.1
uuid 